refactor(emoji): split emoji and spaceEmoji repositories, move emoji/legacyEmoji to core/interfaces
#AT-69
