+++ src/api/api.ts
@@ -27,6 +27,7 @@ import * as spaceIntegrationsRepository from './repositories/spaceIntegrationsRe
 import * as vibeRepository from './repositories/vibeRepository';
 import * as textChatRepository from './repositories/textChatRepository';
 import * as emojiRepository from './repositories/emojiRepository';
+import * as spaceEmojiRepository from './repositories/spaceEmojiRepository';
 import * as worldBuilderRepository from './repositories/worldBuilderRepository';
 
 /**
@@ -69,6 +70,7 @@ export const api = {
   vibeRepository,
   textChatRepository,
   emojiRepository,
+  spaceEmojiRepository,
   worldBuilderRepository
 };
 
@@ -100,5 +102,6 @@ export * from './repositories/spaceTypeRepository/spaceTypeRepository.api.types'
 export * from './repositories/vibeRepository/vibeRepository.api.types';
 export * from './repositories/textChatRepository/textChatRepository.api.types';
 export * from './repositories/emojiRepository/emojiRepository.api.types';
+export * from './repositories/spaceEmojiRepository/spaceEmojiRepository.api.types';
 export * from './repositories/worldBuilderRepository/worldBuilderRepository.api.types';
 export * from './repositories/attendeesRepository/attendeesRepository.api.types';
