+++ src/api/repositories/emojiRepository/emojiRepository.api.types.ts
@@ -1,52 +1,3 @@
-export interface WorldEmojiesRequest {
-  worldId: string;
-}
-
-interface EmojiConfigItemLegacyInterface {
-  emoji: {
-    id: {
-      type: string;
-      data: Buffer;
-    };
-    code: string;
-    hash: string;
-    name: string;
-    createdAt: string | null;
-    updatedAt: string | null;
-  };
-  emojiId: {
-    type: string;
-    data: Buffer;
-  };
-  spaceId: {
-    type: string;
-    data: Buffer;
-  };
-  order: number;
-}
-export interface WorldEmojiesLegacyResponse extends Array<EmojiConfigItemLegacyInterface> {}
-
-interface EmojiConfigItemInterface {
-  id: string;
-  code: string;
-  hash: string;
-  name: string;
-  createdAt: string | null;
-  updatedAt: string | null;
-  spaceId: string;
-  spaceName?: string;
-  order: number;
-}
-export interface WorldEmojiesResponse extends Array<EmojiConfigItemInterface> {}
-
-export interface SpaceEmojiesRequest {
-  spaceId: string;
-}
-
-export interface SpaceEmojiesLegacyResponse extends Array<EmojiConfigItemLegacyInterface> {}
-
-export interface SpaceEmojiesResponse extends Array<EmojiConfigItemInterface> {}
-
 export interface UploadEmojiRequest {
   spaceId: string;
   file: File;
@@ -58,13 +9,6 @@ export interface UploadEmojiResponse {
   emojiId: string;
   message: string;
 }
-
-export interface AssignEmojiToSpaceRequest {
-  emojiId: string;
-  spaceId: string;
-}
-
-export interface AssignEmojiToSpaceResponse extends Array<EmojiConfigItemInterface> {}
 
 export interface DeleteEmojiRequest {
   emojiId: string;
