+++ src/api/repositories/index.ts
@@ -22,3 +22,4 @@ export * from './textChatRepository';
 export * from './worldBuilderRepository';
 export * from './attendeesRepository';
 export * from './emojiRepository';
+export * from './spaceEmojiRepository';
