+++ src/api/repositories/spaceEmojiRepository/index.ts
@@ -1,0 +1,1 @@
+export * from './spaceEmojiRepository.api';
