+++ src/core/interfaces/index.ts
@@ -10,3 +10,4 @@ export * from './stageModePopupOptions.interface';
 export * from './message.interface';
 export * from './coordination.interface';
 export * from './tokenRulesColumnHeader.interface';
+export * from './emoji.interface';
